requirements.txt
weaviate-client
langchain
gptrim
lxml
bs4
python-dotenv
atlassian-python-api
markdownify
sentence-transformers